+++ src/index.html
@@ -45,9 +45,13 @@
             <div class="movies-banners">
             </div>
         </div>
+
     </main>
+    <footer>
+        <p>Create by Microverse CC license</p>
+    </footer>
     <div id="popup-modal"></div>
-    
+
 </body>
 
 </html>
